requirements.txt
# SOVRA Python Dependencies
# ============================================

# --- Core ---
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
httpx>=0.27.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# --- RAG / Memory ---
chromadb>=0.5.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# --- Embeddings ---
sentence-transformers>=3.0.0

# --- Self-Evolution (LoRA) ---
# Note: For CPU-only, use the standard versions
# For GPU, install unsloth separately: pip install unsloth
torch>=2.1.0
transformers>=4.45.0
peft>=0.13.0
datasets>=3.0.0
trl>=0.12.0
accelerate>=1.0.0

# --- Scheduler ---
apscheduler>=3.10.0

# --- Utilities ---
rich>=13.0.0
click>=8.1.0
psutil>=5.9.0
